% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_demand.R
\name{region_demand}
\alias{region_demand}
\title{Calculate demand for each region}
\usage{
region_demand(destinations, origins, demand, population, n_iter)
}
\arguments{
\item{destinations}{Dataframe of the destinations of each point of demand}

\item{origins}{Dataframe of the origins of each point of supply}

\item{demand}{Demand associated with each destination region}

\item{population}{Population counts in each destination subregion}

\item{n_iter}{Number of iterations to run the bootstrapping algorithm}
}
\value{
A list containing a dataframe with the access amounts and a dataframe with each
iteration's demand.
}
\description{
Calculate demand for each region
}
