\name{safc-package}
\alias{safc-package}
\alias{safc}
\docType{package}
\title{
\packageTitle{safc}
}
\description{
\packageDescription{safc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{safc}
\packageIndices{safc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{safc}

Maintainer: \packageMaintainer{safc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
