% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safc-package.R
\docType{package}
\name{safc}
\alias{safc}
\alias{safc-package}
\title{safc}
\description{
Functions For Calculating Spatially Adaptive Floating Catchments
}
\author{
David DeLara <sfq1@cdc.gov>
}
\keyword{internal}
