% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safc_availability.R
\name{safc_availability}
\alias{safc_availability}
\title{Calculate demand for each region}
\usage{
safc_availability(
  supply,
  demand,
  region,
  distmat = NULL,
  max_dist = Inf,
  n_boot = NULL
)
}
\arguments{
\item{supply}{Tibble with origin matrix and supply for each origin}

\item{demand}{Tibble with destination matrix and demand for each destination. May contain
optional bootstrapping information.}

\item{region}{Tibble with information to aggregate from destination to region level}

\item{distmat}{Optional distance matrix. If \code{distmat = NULL}, a distance matrix will be
generated automatically using the origin/destination information}

\item{max_dist}{A double describing the cutoff distance in the distance matrix calculation}

\item{n_boot}{Number of iterations to run the bootstrapping algorithm}
}
\value{
A list containing model information and the estimated demand met for each region
}
\description{
Calculate demand for each region
}
